Add correct relative -> absolute path converter
For the :Rdebugger command, added the relative -> absolute path
converter. E.g. command :Rdebugger 'rackup config.ru' will be converted
to "'/usr/bin/rackup' config.ru". First, it tries to find place of the
script by 'which' command, if it is unsuccessful, then it assumes it is
relative path with root dir = current dir.

It also correctly works with spaces in directories, so Issue #37 is
fixed.