Fixed logging - it is strongly recommended to take the latest!
Now it logs only if g:ruby_debugger_debug_mode = 1. Otherwise, it
doesn't write anything to the log file. It actually affects performance
very very much, due to stupid implemenation of logging - to add one line
to the log file, it reads whole the file to memory, add one line there,
and then write whole file contents to the file back. Of course when the
file is large, it will work very slow.

So, enable debugging mode only if you have any issues with the debugger.
And please pull the latest and update the plugin in your ~/.vim dir,
because now it works much faster.